fix: PR feedback
