--- packages/jazz-react-native-auth-betterauth/src/index.ts
@@ -1,9 +1,0 @@
-// Re-export types and utilities from jazz-react-auth-betterauth source files
-export type {
-  AuthClient,
-  Session,
-  User,
-} from "jazz-react-auth-betterauth";
-export { SocialProvider } from "jazz-react-auth-betterauth/src/lib/social";
-export { AuthProvider, useAuth } from "./contexts/Auth.js";
-export { useBetterAuth } from "./hooks.js";
