--- packages/jazz-react-native-auth-betterauth/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2020",
-    "module": "ESNext",
-    "moduleResolution": "bundler",
-    "esModuleInterop": true,
-    "strict": true,
-    "skipLibCheck": true,
-    "declaration": true,
-    "jsx": "react-jsx",
-    "outDir": "dist",
-    "allowJs": true,
-    "resolveJsonModule": true
-  },
-  "include": ["src"]
-}
